refact: ajusta mapeamento JPA e limpa propriedades de banco
- Atualiza entidades e queries nativas de Participante para coletor.

- Remove propriedades com senhas vazadas no application-mysql.properties.

- Ajusta inicialização de conexões e constraints.

+++ src/main/java/br/unesp/fct/evcomp/controller/AtividadeController.java
@@ -37,7 +37,7 @@ public class AtividadeController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<?> buscarAtividadeWeb(@PathVariable Long id) {
+    public ResponseEntity<?> buscarAtividadeWeb(@PathVariable Integer id) {
         Optional<Atividade> atividade = atividadeRepository.findById(id);
         if (atividade.isPresent()) {
             return ResponseEntity.ok(atividade.get());
