+++ src/main/java/br/unesp/fct/evcomp/repository/ColetorRepository.java
@@ -3,7 +3,7 @@ package br.unesp.fct.evcomp.repository;
 import br.unesp.fct.evcomp.domain.ColetorDePresenca;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ColetorRepository extends JpaRepository<ColetorDePresenca, Long> {
+public interface ColetorRepository extends JpaRepository<ColetorDePresenca, Integer> {
     default boolean verificarColetor(String participanteId) { return false; }
     default void atribuirPapelColetor(String participanteId) { }
     default boolean deletarColetor(String eventoId, String coletorId) { return false; }
